% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropbraces}
\alias{dropbraces}
\title{Bain lúibíní slabhracha}
\usage{
dropbraces(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropbraces}{Bain lúibíní slabhracha as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain lúibíní slabhracha
}
\examples{
x <- dropbraces(x)
x <- x \%>\% dplyr::mutate(col_name=dropbraces(col_name))
}
