% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R, R/hello.R
\name{dropbrackets}
\alias{dropbrackets}
\title{Drop brackets}
\usage{
dropbrackets(x)

dropbrackets(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropbrackets}{Drop brackets from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop brackets

Drop Brackets
}
\examples{
x <- dropbrackets(x)
x <- x \%>\% dplyr::mutate(col_name=dropbrackets(col_name))
x <- dropbrackets(x)
x <- x \%>\% dplyr::mutate(col_name=dropbrackets(col_name))
}
