% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropcolon}
\alias{dropcolon}
\title{Bain idirstadanna}
\usage{
dropcolon(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropcolon}{Bain idirstadanna as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain idirstadanna
}
\examples{
x <- dropcolon(x)
x <- x \%>\% dplyr::mutate(col_name=dropcolon(col_name))
}
