% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropcomma}
\alias{dropcomma}
\title{Drop commas}
\usage{
dropcomma(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropcomma}{Drop commas from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop commas
}
\examples{
x <- dropcomma(x)
x <- x \%>\% dplyr::mutate(col_name=dropcomma(col_name))
}
