% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropnacol}
\alias{dropnacol}
\title{bain colúin NA}
\usage{
dropnacol(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropnacol}{bain colúin ina bhfuil luachanna NA}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain colúin NA
}
\examples{
x <- dropnacol(x)
}
