% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropparen}
\alias{dropparen}
\title{Drop parentheses}
\usage{
dropparen(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropparen}{Drop parentheses from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop parentheses
}
\examples{
x <- dropparen(x)
x <- x \%>\% dplyr::mutate(col_name=dropparen(col_name))
}
