% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropquotes}
\alias{dropquotes}
\title{bain comhartha athfhriotail}
\usage{
dropquotes(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropquotes}{bain comhartha athfhriotail as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain comhartha athfhriotail
}
\examples{
x <- dropquotes(x)
x <- x \%>\% dplyr::mutate(col_name=dropquotes(col_name))
}
