% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropsemi}
\alias{dropsemi}
\title{Drop semicolons}
\usage{
dropsemi(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropsemi}{Drop semicolons from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop semicolons
}
\examples{
x <- dropsemi(x)
x <- x \%>\% dplyr::mutate(col_name=dropsemi(col_name))
}
