% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropspaces}
\alias{dropspaces}
\title{bain spásanna}
\usage{
dropspaces(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropspaces}{bain spásanna as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain spásanna
}
\examples{
x <- dropspaces(x)
x <- x \%>\% dplyr::mutate(col_name=dropspaces(col_name))
}
