% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropstop}
\alias{dropstop}
\title{Drop full stops}
\usage{
dropstop(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropstop}{Drop full stops from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop full stops
}
\examples{
x <- dropstop(x)
x <- x \%>\% dplyr::mutate(col_name=dropstop(col_name))
}
