% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{droptext}
\alias{droptext}
\title{bain carachtair nó teaghráin atá i liosta}
\usage{
droptext(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{droptext}{bain carachtair nó teaghráin atá i liosta as tacar sonraí}

\item{todrop}{liosta carachtar nó teaghrán le baint as an tacar sonraí: todrop <- paste(c("value1", "value2"...), collapse = "|")}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain carachtair nó teaghráin atá i liosta
}
\examples{
x <- droptext(x)
}
