% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{firstrow}
\alias{firstrow}
\title{athraigh na chéad sraitheanna}
\usage{
firstrow(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{firstrow}{athraigh luachanna an chéad ró go hainmneacha na gcolún}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
athraigh na chéad sraitheanna
}
\examples{
x <- firstrow(x)
}
