% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{lowerall}
\alias{lowerall}
\title{Change to lowercase}
\usage{
lowerall(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{lowerall}{Change characters to lowercase}
}
\value{
Changes are saved to the dataset
}
\description{
Change to lowercase
}
\examples{
x <- lowerall(x)
x <- x \%>\% dplyr::mutate(col_name=lowerall(col_name))
}
