% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{pcdec}
\alias{pcdec}
\title{Calculate percentage of total (decimal)}
\usage{
pcdec(x, pcdec, existing_col_name)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{pcdec}{New column name to be created}

\item{existing_col_name}{Existing column name to call}
}
\value{
Changes are saved to the dataset
}
\description{
Calculate percentage of total (decimal)
}
\examples{
x <- pcdec(x, new_col_name, existing_col_name)
}
