% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{pclab}
\alias{pclab}
\title{Cruthaigh lipéid céatadáin ó luachanna deachúlacha}
\usage{
pclab(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{pclab}{Lipéid chéatadáin a chruthú le haghaidh plota ó luachanna deachúlacha}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Cruthaigh lipéid céatadáin ó luachanna deachúlacha
}
\examples{
x <- pclab(x)
}
