% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{upperall}
\alias{upperall}
\title{Athraigh go litreacha móra}
\usage{
upperall(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{upperall}{Athraigh an téacs go litreacha móra}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Athraigh go litreacha móra
}
\examples{
x <- upperall(x)
x <- x \%>\% dplyr::mutate(col_name=upperall(col_name))
}
