% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropbraces}
\alias{dropbraces}
\title{bain lúibíní slabhracha}
\usage{
dropbraces(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropbraces}{bain lúibíní slabhracha as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain lúibíní slabhracha
}
\examples{
x <- dropbraces(x)
x <- x \%>\% dplyr::mutate(col_name=dropbraces(col_name))
}
