% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropbraces}
\alias{dropbraces}
\title{Drop braces}
\usage{
dropbraces(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropbraces}{Drop braces from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop braces
}
\examples{
x <- dropbraces(x)
x <- x \%>\% dplyr::mutate(col_name=dropbraces(col_name))
}
