% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R, R/hello.R
\name{dropbrackets}
\alias{dropbrackets}
\title{bain lúibíní cearnacha}
\usage{
dropbrackets(x)

dropbrackets(x)
}
\arguments{
\item{x}{}

\item{dropbrackets}{bain lúibíní cearnacha as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain lúibíní cearnacha

Drop Brackets
}
\examples{
x <- dropbrackets(x)
x <- x \%>\% dplyr::mutate(col_name=dropbrackets(col_name))
x <- dropbrackets(x)
x <- x \%>\% dplyr::mutate(col_name=dropbrackets(col_name))
}
