% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropcolon}
\alias{dropcolon}
\title{bain idirstadanna}
\usage{
dropcolon(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropcolon}{bain idirstadanna as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain idirstadanna
}
\examples{
x <- dropcolon(x)
x <- x \%>\% dplyr::mutate(col_name=dropcolon(col_name))
}
