% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropcolon}
\alias{dropcolon}
\title{Drop colons}
\usage{
dropcolon(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropcolon}{Drop colons from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop colons
}
\examples{
x <- dropcolon(x)
x <- x \%>\% dplyr::mutate(col_name=dropcolon(col_name))
}
