% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropcomma}
\alias{dropcomma}
\title{Bain camóga}
\usage{
dropcomma(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropcomma}{Bain camóga as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain camóga
}
\examples{
x <- dropcomma(x)
x <- x \%>\% dplyr::mutate(col_name=dropcomma(col_name))
}
