% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropparen}
\alias{dropparen}
\title{Bain lúibíní}
\usage{
dropparen(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropparen}{Bain lúibíní as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain lúibíní
}
\examples{
x <- dropparen(x)
x <- x \%>\% dplyr::mutate(col_name=dropparen(col_name))
}
