% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropparen}
\alias{dropparen}
\title{bain lúibíní}
\usage{
dropparen(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropparen}{bain lúibíní as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain lúibíní
}
\examples{
x <- dropparen(x)
x <- x \%>\% dplyr::mutate(col_name=dropparen(col_name))
}
