% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropquotes}
\alias{dropquotes}
\title{Drop quotation marks}
\usage{
dropquotes(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{dropquotes}{Drop quotation marks from a dataset}
}
\value{
Changes are saved to the dataset
}
\description{
Drop quotation marks
}
\examples{
x <- dropquotes(x)
x <- x \%>\% dplyr::mutate(col_name=dropquotes(col_name))
}
