% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropquotes}
\alias{dropquotes}
\title{Bain comhartha athfhriotail}
\usage{
dropquotes(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropquotes}{Bain comhartha athfhriotail as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain comhartha athfhriotail
}
\examples{
x <- dropquotes(x)
x <- x \%>\% dplyr::mutate(col_name=dropquotes(col_name))
}
