% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropsemi}
\alias{dropsemi}
\title{bain leathstadanna}
\usage{
dropsemi(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropsemi}{bain leathstadanna as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain leathstadanna
}
\examples{
x <- dropsemi(x)
x <- x \%>\% dplyr::mutate(col_name=dropsemi(col_name))
}
