% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropsemi}
\alias{dropsemi}
\title{Bain leathstadanna}
\usage{
dropsemi(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropsemi}{Bain leathstadanna as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain leathstadanna
}
\examples{
x <- dropsemi(x)
x <- x \%>\% dplyr::mutate(col_name=dropsemi(col_name))
}
