% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropspaces}
\alias{dropspaces}
\title{Bain spásanna}
\usage{
dropspaces(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropspaces}{Bain spásanna as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain spásanna
}
\examples{
x <- dropspaces(x)
x <- x \%>\% dplyr::mutate(col_name=dropspaces(col_name))
}
