% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropstop}
\alias{dropstop}
\title{bain lánstadanna}
\usage{
dropstop(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropstop}{bain lánstadanna as tacar sonraí}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
bain lánstadanna
}
\examples{
x <- dropstop(x)
x <- x \%>\% dplyr::mutate(col_name=dropstop(col_name))
}
