% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{dropstop}
\alias{dropstop}
\title{Bain lánstadanna}
\usage{
dropstop(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{dropstop}{Bain lánstadanna as tacar sonraí}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain lánstadanna
}
\examples{
x <- dropstop(x)
x <- x \%>\% dplyr::mutate(col_name=dropstop(col_name))
}
