% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{droptext}
\alias{droptext}
\title{Drop characters or strings contained in a list}
\usage{
droptext(x)
}
\arguments{
\item{x}{Dataset to which the function is applied}

\item{droptext}{Drop characters or strings contained in a list from a dataset}

\item{todrop}{List of characters or strings to drop from the dataset. Create using: todrop <- paste(c("value1", "value2"...), collapse = "|")}
}
\value{
Changes are saved to the dataset
}
\description{
Drop characters or strings contained in a list
}
\examples{
x <- droptext(x)
}
