% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{droptext}
\alias{droptext}
\title{Bain carachtair nó teaghráin atá i liosta}
\usage{
droptext(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{droptext}{Bain carachtair nó teaghráin atá i liosta as tacar sonraí}

\item{todrop}{Liosta carachtar nó teaghrán le baint as an tacar sonraí: todrop <- paste(c("value1", "value2"...), collapse = "|")}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Bain carachtair nó teaghráin atá i liosta
}
\examples{
x <- droptext(x)
}
