% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{lowerall}
\alias{lowerall}
\title{Athraigh go litreacha beaga}
\usage{
lowerall(x)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{lowerall}{Athraigh an téacs go litreacha beaga}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Athraigh go litreacha beaga
}
\examples{
x <- lowerall(x)
x <- x \%>\% dplyr::mutate(col_name=lowerall(col_name))
}
