% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{lowerall}
\alias{lowerall}
\title{athraigh go litreacha beaga}
\usage{
lowerall(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{lowerall}{athraigh an téacs go litreacha beaga}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
athraigh go litreacha beaga
}
\examples{
x <- lowerall(x)
x <- x \%>\% dplyr::mutate(col_name=lowerall(col_name))
}
