% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{pcdec}
\alias{pcdec}
\title{Ríomh céatadán den iomlán (deachúil)}
\usage{
pcdec(x, pcdec, existing_col_name)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{pcdec}{New column name to be created}

\item{existing_col_name}{Existing column name to call}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Ríomh céatadán den iomlán (deachúil)
}
\examples{
x <- pcdec(x, new_col_name, existing_col_name)
}
