% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{pcdec}
\alias{pcdec}
\title{ríomh céatadán den iomlán (deachúil)}
\usage{
pcdec(x, pcdec, existing_col_name)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{pcdec}{ainm nua colúin le cruthú}

\item{existing_col_name}{ainm colúin atá ann cheana féin le glaoch}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
ríomh céatadán den iomlán (deachúil)
}
\examples{
x <- pcdec(x, new_col_name, existing_col_name)
}
