% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{pclab}
\alias{pclab}
\title{cruthaigh lipéid céatadáin ó luachanna deachúlacha}
\usage{
pclab(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{pclab}{lipéid chéatadáin a chruthú le haghaidh plota ó luachanna deachúlacha}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
cruthaigh lipéid céatadáin ó luachanna deachúlacha
}
\examples{
x <- pclab(x)
}
