% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{pclab}
\alias{pclab}
\title{Create percentage labels from decimal values}
\usage{
pclab(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{pclab}{Create percentage labels for plotting from decimal values}
}
\value{
Changes are saved to the dataset
}
\description{
Create percentage labels from decimal values
}
\examples{
x <- pclab(x)
}
