% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{percent}
\alias{percent}
\title{Ríomh céatadán den iomlán (deachúil cruinn)}
\usage{
percent(x, percent, existing_col_name)
}
\arguments{
\item{x}{An tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{percent}{New column name to be created}

\item{existing_col_name}{Existing column name to be called}
}
\value{
Sábháiltear athruithe sa tacar sonraí
}
\description{
Ríomh céatadán den iomlán (deachúil cruinn)
}
\examples{
x <- percent(x, new_col_name, existing_col_name)
}
