% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{percent}
\alias{percent}
\title{Calculate percentage of total (rounded decimal)}
\usage{
percent(x, percent, existing_col_name)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{percent}{New column name to be created}

\item{existing_col_name}{Existing column name to be called}
}
\value{
Changes are saved to the dataset
}
\description{
Calculate percentage of total (rounded decimal)
}
\examples{
x <- percent(x, new_col_name, existing_col_name)
}
