% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{percent}
\alias{percent}
\title{ríomh céatadán den iomlán (deachúil cruinn)}
\usage{
percent(x, percent, existing_col_name)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{percent}{ainm nua colúin le cruthú}

\item{existing_col_name}{ainm colúin atá ann cheana féin le glaoch}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
ríomh céatadán den iomlán (deachúil cruinn)
}
\examples{
x <- percent(x, new_col_name, existing_col_name)
}
