% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{upperall}
\alias{upperall}
\title{athraigh go litreacha móra}
\usage{
upperall(x)
}
\arguments{
\item{x}{an tacar sonraí ar a gcuirtear an fheidhm i bhfeidhm}

\item{upperall}{athraigh an téacs go litreacha móra}
}
\value{
sábháiltear athruithe sa tacar sonraí
}
\description{
athraigh go litreacha móra
}
\examples{
x <- upperall(x)
x <- x \%>\% dplyr::mutate(col_name=upperall(col_name))
}
