% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glan.R
\name{upperall}
\alias{upperall}
\title{Change to uppercase}
\usage{
upperall(x)
}
\arguments{
\item{x}{Dataset to which function is applied}

\item{upperall}{Change characters to uppercase}
}
\value{
Changes are saved to the dataset
}
\description{
Change to uppercase
}
\examples{
x <- upperall(x)
x <- x \%>\% dplyr::mutate(col_name=upperall(col_name))
}
